Combined store implementations for simplicity (no cost in production)
