+++ package.json
@@ -6,7 +6,7 @@
     "clean": "rimraf dist",
     "webpack": "cross-env NODE_ENV=production webpack -p --config webpack.config.prod.js",
     "assets": "copy static/* dist && copy favicon.ico dist && copy CNAME dist",
-    "build": "run-s clean webpack assets",
+    "build:static": "run-s clean webpack assets",
     "start": "node devServer.js",
     "lint:js": "eslint src --ext .js --ext .jsx",
     "lint": "run-s lint:*",
@@ -14,15 +14,15 @@
     "critical": "critical ./dist/index.html --inline --base ./dist -H ./dist/index.html > /dev/null",
     "surge": "surge -p dist -d literate-army.surge.sh",
     "gh-pages": "gh-pages -d dist",
-    "deploy": "run-s build critical gh-pages",
+    "deploy": "run-s build:static critical gh-pages",
     "precommit": "run-s lint test"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/tsaiDavid/simple-redux-boilerplate.git"
+    "url": "git@github.com:csb324/holyfucktheelectionistomorrow.git"
   },
   "license": "MIT",
-  "homepage": "https://github.com/tsaiDavid/simple-redux-boilerplate",
+  "homepage": "https://github.com/csb324/holyfucktheelectionistomorrow",
   "devDependencies": {
     "autoprefixer": "^6.5.2",
     "babel-core": "^6.3.15",
@@ -38,7 +38,7 @@
     "cross-env": "^3.1.3",
     "css-loader": "^0.26.0",
     "dotenv": "^2.0.0",
-    "eslint": "^3.10.2",
+    "eslint": "^3.11.0",
     "eslint-config-airbnb": "^13.0.0",
     "eslint-plugin-babel": "^4.0.0",
     "eslint-plugin-import": "^2.2.0",
